[issue-660][feature]: Refactor internal-api module to have less dependecies on 3d party libraries
