+++ internal-api/src/main/scala/com/griddynamics/genesis/plugin/StepCoordinatorFactory.scala
@@ -23,7 +23,6 @@
 package com.griddynamics.genesis.plugin
 
 import com.griddynamics.genesis.workflow.{StepCoordinator, Step}
-import com.griddynamics.genesis.util.Logging
 
 trait StepCoordinatorFactory {
     def apply(step: Step, context: StepExecutionContext): StepCoordinator
@@ -33,24 +32,3 @@ trait PartialStepCoordinatorFactory extends StepCoordinatorFactory {
     def isDefinedAt(step: Step): Boolean
 }
 
-class CompositeStepCoordinatorFactory(factories: Array[PartialStepCoordinatorFactory])
-    extends StepCoordinatorFactory with Logging{
-
-    def apply(step: Step, context: StepExecutionContext) = {
-        factories.find(_.isDefinedAt(step)) match {
-            case Some(factory) => try {
-                factory.apply(step, context)
-            } catch {
-                case e: Throwable => {
-                    log.error(e, "Failed to create step coordinator for %s".format(step))
-                    throw new RuntimeException("Failed to start step %s due to error: %s".format(step, e.getCause match {
-                        case null => e.getMessage
-                        case _ => e.getCause.getMessage
-                    }), e)
-                }
-            }
-            case None => throw new RuntimeException("Failed to find coordinator for '%s'".
-                format(step))
-        }
-    }
-}
