+++ internal-api/src/main/scala/com/griddynamics/genesis/util/Logging.scala
@@ -1,4 +1,4 @@
-/**
+/*
  * Copyright (c) 2010-2012 Grid Dynamics Consulting Services, Inc, All Rights Reserved
  *   http://www.griddynamics.com
  *
