+++ local-users/pom.xml
@@ -49,6 +49,11 @@
         </dependency>
 
         <dependency>
+            <groupId>com.griddynamics.genesis</groupId>
+            <artifactId>util</artifactId>
+        </dependency>
+
+        <dependency>
             <groupId>org.squeryl</groupId>
             <artifactId>${org.squeryl.id}</artifactId>
         </dependency>
