+++ nexus/pom.xml
@@ -44,6 +44,10 @@
             <artifactId>internal-api</artifactId>
         </dependency>
         <dependency>
+            <groupId>com.griddynamics.genesis</groupId>
+            <artifactId>util</artifactId>
+        </dependency>
+        <dependency>
             <groupId>commons-codec</groupId>
             <artifactId>commons-codec</artifactId>
         </dependency>
