+++ notification/pom.xml
@@ -25,6 +25,16 @@
         </dependency>
 
         <dependency>
+            <groupId>com.griddynamics.genesis</groupId>
+            <artifactId>util</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-context-support</artifactId>
+        </dependency>
+
+        <dependency>
             <groupId>javax.mail</groupId>
             <artifactId>mail</artifactId>
             <version>1.4.5</version>
