+++ util/src/main/scala/com/griddynamics/genesis/util/Closeables.scala
