+++ util/src/main/scala/com/griddynamics/genesis/util/InputUtil.scala
