+++ util/src/main/scala/com/griddynamics/genesis/util/SafeOperation.scala
