+++ util/src/main/scala/com/griddynamics/genesis/util/ScalaCaseClassSigParser.scala
