+++ util/src/main/scala/com/griddynamics/genesis/util/ScalaUtils.scala
