+++ util/src/main/scala/com/griddynamics/genesis/util/Shell.scala
