+++ util/src/main/scala/com/griddynamics/genesis/util/TryingUtil.scala
